requirements_gradio.txt
gradio==4.16.0
ultralytics==8.3.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0

